rename: UserInfo를 application 패키지로 이동
