+++ src/main/java/team/hackerping/nanuri/user/presentation/UserRestController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import team.hackerping.nanuri.user.application.UserService;
-import team.hackerping.nanuri.user.domain.UserInfo;
+import team.hackerping.nanuri.user.application.UserInfo;
 import team.hackerping.nanuri.user.presentation.dto.UserResponse;
 
 @RequiredArgsConstructor
